chore: extract field templates view
